refactor: migrate bullet angles from 256-degree to float radians
Store bullet/laser entity angles as double radians instead of uint8_t deg256, eliminating the atan8 quantization and per-frame grid snap in homing. Command structs stay uint8_t (ECL script boundary); MAP.PAK and script_tool untouched. Demos/replays will desync.

+++ GIAN07/core/point.h
@@ -9,5 +9,5 @@
 // Structure for managing angular coordinates
 struct DegPoint {
   int x, y;  // Coordinates
-  uint8_t d; // Angle
+  double d;  // Angle (radians)
 };
